WebExtension: use WebSocket to pass intercepted downloads when applicable
When the dl-mngr application is already running, it is faster to use
a WebSocket connection to communicate with it.
When WebSocket port is unknown or invalid, intercepted downloads are
passed through the native application, which asks for the WebSocket
port and returns it. Then next requests can use it (and fallback to
native application upon issue).