Better handling of application issues
App: when an uncaught exception happens, the logged stderr messages
may not be fully sent before nodejs exits. So handle those to
initiate a proper shutdown ensuring pending messages have been sent.
The uncaught exception is also sent to the extension.

App: better code comments on the implemented protocol

WebExtension: handle application messages needing manual clearing
When the native application encounters an uncaught exception, it is
notified to the extension and needs user attention: the message is
kept visible in the browser action (and also settings) page until
cleared; a visual hint (badge) is set on the icon in this case.
Similarly, notify if we receive an error (explicit or non-0 code)
from the remote application (through the spawned child process in
the native application, or a WebSocket message).

WebExtension: also notify upon action timeout