WebExtension: don't intercept request for what may be page content
Some pages don't have a content length in the received response.
Text and images are already excluded from interception in this case.
We may still encounter pages that don't have a proper content type,
so try to determine if the content may actually be a standard page
to display as opposed to download: determine if the URL belongs to
a page or a file, taking ino account known page extensions.