WebExtension: erase download entry more properly
Delay nominal erasing because when the download is triggered by
another extension, that extension may fail to properly take into
account the download cancellation if we erase it too fast, which
means those extensions need the downloads entry to still exist.
Delaying by 2s appears to be enough.

On Firefox (v66) 'browser.downloads.erase' removes the entry from
the downloads list, but - at least for downloads triggered by an
extension - an entry still remains in the downloads history.
'browser.browsingData.removeDownloads' may be able to cleanup there
too, but it does not appear to work as intended maybe because
removing extensions data is not yet supported.
Fortunately, 'browser.history.deleteUrl' does the trick.