Filter out undefined (and null) download message fields
If a field is undefined, it is correctly ignored when passed to
dl-mngr (through WebSocket or native application call).
If a field was null, it was also correctly ignored when passed
through WebSocket, but the native application call would translate
it to the "null" string.

WebExtension: filter out undefined/null fields in message before
passing it (through WebSocket or native application call).

App: also consider null values as undefined ones.