WebExtension: filter internal messages by target
Messages can be sent between components of the extension.
If a component sends a message, all the other components currently
running will receive it. Since we also log unhandled messages in
each component, add a simple target filtering system so that only
concerned components will actually handle the received message.

It is done in the messages handling generic code, where we check
whether a received message target matches the current component.
It remains possible to send a message to all components, or for a
component to handle all messages whatever the requested target.