WebExtension: button to ignore next interception
Added a button in the 'browser action' (and 'options') page that
allows to ignore the next interception - request or download.
Interceptions are re-enabled when either the next interception is
triggered and ignored, or after 20s if none is triggered.
Using the button again before interception is triggered will cancel
ignoring the next interception.