webext-native-messaging: dedicated dl-mngr client object
To allow other features to trigger downloads, move dl-mngr client
code (relying on native app and websocket) in a dedicated class.
One instance is created in the background page and used when
needed.

Cleanup/simplify a bit the code to use async/await instead of
Promise chaining.

Let the client code add cookie, user agent and comment when
needed.