webext-native-messaging: refactor messaging handling
Instead of 'broadcasting' through browser.runtime.sendMessage, use
dedicated Ports and have background script manage forwarding when
applicable.

This allows to selectively post messages to actual targets, and also
prevent errors when actual target is not listening(/running).

Share code with native application messaging, which uses Port.
As for native application messaging, correlation is needed to pair
requests and replies (while runtime.sendMessage does it implicitly
by returning a Promise with the receiver response).

Fallback to broadcast if Port is not connected, even though it should
not be needed/useful.