webext-native-messaging: better handle settings value update
When setting is removed, the new value is undefined.
To handle this case, remember the default value given upon
creating the setting, and use it here.

Also do notification during value update when applicable.
This way it is not done if the default value matches the
previous value when removing a setting.