native-messaging: handle fallback value for setting kinds
Settings can be changed through forms. Without special handling,
we actually cannot (and do not need to) use undefined/null values.
So for each kind of setting, use a fallback value when given
default value is undefined/null:
  - boolean: false
  - int: 0
  - string: ''

Fixes the current string settings for which the default value was
given as undefined: the HTLM field fails when given undefined/null
which required us to set '' instead, which in turn would make this
empty string uselessly saved in storage.