native-messaging: only update browser action videos for focused window
When the browser action page is opened, videos added to the active
tab of another window would be notified to the action page.

Ensure that only the sources related to the focused window are
notified, since only the browser action page of the focused window
can be active and listening.