Code cleanup and minor refactoring
Have a more consistent usage of undefined, null and '' in the code.
e.g. we usually need to check if a string is defined and non-empty, in
which case 'if (s)' is enough.
The same applies to object instances.

We usually don't have to define object fields as undefined, but can
simply not declare them, delete them when needed, and 'if (obj.field)'
to check truthy fields. In which case we can also 'if ("field" in obj)'
to check if a field is present (even if its value is undefined).

Not 'return'ing anything in a function automatically returns 'undefined'
which can simplify some function code when needing to return an object
field value only when the object exists.

To test numbers, Number.isInteger (integer), Number.isFinite (number)
or Number.isNaN can also be used.

Ensure a tabId is passes as tab successor, which simplifies code.