Refactor content script execution
Merge all (three for now) content scripts into one: instead of deciding
in which frames to inject each one, inject it in all frames and decide
in each frame which features are needed.
This reduces the size of the extension, and overall content scripts,
since many functions and classes are common to all scripts, and don't
have to be duplicated in multiple bundles anymore.
Reducing the number of scripts may also benefit execution times.

Instead of trying to actively manage content scripts injection, let
the browser do it nominally by declaring the bundle in the manifest.
Then have the content script generate one UUID per frame, communicated
to the background script ASAP, and added to all sent messages.
This is the new way to detect new frames - instead of listening to
webNavigation stages - which allows to discover them a bit earlier
while greatly simplifying frames management in tabs handler: no more
need to inject script and check result etc.
We assume that most of the time the ordering of events will be correct.

In order for content scripts to properly work, the background script
now has to listen to messages/ports creation immediately (not after
waiting for settings asynchronously): if not, content scripts may fail
to properly register on the background side when starting.
This also concerns a few other resources in the background script that
are needed when handling messages received from content scripts.
For proper and more consistent handling, all background resources are
created immediately, while some of them internally wait for settings
before actually being fully initialized and running.