Refactor code execution helper
Pass a map of parameters, which allows more flexibility.
Share more code when code is executed only once.
Have caller rely on the settings getKey method to use setting
for both generating a cache key and get the script code.

Minor cleanup in settings to declare them in alphabetical order.