Enhanced video filename refining
Give complete control to the executed script.
There is no more nominal sanitizing done on title: it's up to the
script to do whatever is necessary.
Full access to the current naming/title is given to the script, and
many methods can be used to help refining the name/title.