Remove settings with default value from local storage
When changing a setting value, we only write local storage when
value is not the setting default: with default value, we remove
it from storage (if present).
In older versions, that was not always the case: some settings may
be stored while they do have the default value. It could also
happen due to importing settings.

We do want to only keep non-default setting values in storage:
makes upgrading better if we decide to change a setting default
value, as all instances still using default would use the new
value.

Thus, when initializaing settings in background script, remove
those present in storage with default value.
Don't bother trying to filter out settings upon importing: those
imported with default value will be removed next time. We rarely
upgrade version after importing settings, and that would only be
an issue when importing setting with a default value that then
changes after upgrade.