Links catcher: remove legacy/unneeded code
In older versions of Firefox, it was necessary to use 'setCapture' (and
'releaseCapture') so that mouse movements could be tracked even outside
the web page.
Otherwise when the pointer was outside the application window, things
were not working as expected (and we could not refresh caught links).

This appears to not be necessary anymore.
Moreover these methods are not standard (basically only supported by
Firefox), deprecated and can be removed at any time.
A standard equivalent, if needed, is setPointerCapture (and
releasePointerCapture).