tb: fix visual glitches
Use an helper class to override class instance methods.
Makes it easier to setup/reset overridden methods.

In overridden 'refreshItems', we need to rely more on original behaviour
to prevent visual glitches: selecting an event that is not visible in the
calendar view does trigger a refresh, and we must not re-populate the
list in this case.
So always call original method, ensuring it only works on a day for our
use case, and if it did refresh the list, then do our own filtering.
This is the best compromise to stay as close as possible to the original
behaviour.

More generally, better handle invalidation, and do it when appopriate.

Catch more error cases.
Add debug logs.