tb: use available context tools to setup specific window
Instead of iterating over all windows, target the window id indicated by
the background script.
The context has some tools to retrieve the actual real window from a
webextension window id.

Conversely, we can retrieve the (webextension) window id for a given
real window.
So use it in debug logs, instead of logging the window object.

Keep the code preventing to setup a window twice, as belt and
suspenders.