tb: show most appropriate occurrence of recurrent event
When showing all events, for a recurring event actually display one
occurrence (the next or previous one) instead of simply the parent
event.
This makes the date ordering more natural if used.
This also allows to modify the next occurrence or base event if needed.

If next occurrence has been removed, native code usually fails to get
the next one and we get the previous occurrence instead.