ff: refactor menu handling to work with groups
Along pure entries, a group can be added: one or more entries can be
added in a group, and they will appear next to each other in the
resulting menu.
No submenu is created: a group entries will appear at the same level of
other parent entries or other groups.