ff: use JSON when calling dl-mngr through native application
Use the new '--json' option.
Instead of building a CLI with all options extracted from the download
details, just pass them as JSON through its stdin: this is the exact
same JSON we pass through WebSocket to dl-mngr (once the WebSocket port
is known and accessible).
And read back the program response as JSON too.

This has a number of benefits:
 - simplifies code
 - when listing running processes, download details won't appear in the
   process command line; especially for the long-lived first unique
   instance started
 - more complex parameters can be passed as JSON compared to what is
   possible with CLI option parsing
 - bigger data can be passed through stdin compared to CLI
 - passing JSON through CLI allows easier parity with WebSocket usage
 - command output would be able to contain more information if needed