ff: initiate actual video download in one place only
Video download can be triggered from two places:
 - contextual menu entry
 - browser action page (click on element)

Each source would initiate actual download a different way:
 - menu entry is directly related to video source handler, which manages
   it
 - browser action page would pass back download information to
   background script and let it initiate download

Now, let video source handler be the only one to initiate the download:
use a new message kind so that background script passes the information
to the video source handler.
Keep code handling a download initiated by the initial message kind in
the background script: it is only a few lines of code, and may still be
useful to keep around in the future.