ff: handle HLS playlist as video source
Detect master and stream playlists.
Handling is enabled/disabled by a new setting.

Stream playlist (deduced from master, or intercepted as-is) are handled
as proper video source.
The stream playlist content is retrieved and passed to the download
application.
If keys are defined, they are retrieved and passed too.
It is up to the external download application to handle this case:
 - given URL and filename is for the playlist (usually .m3u8)
 - playlist content passed as-is
 - keys, if any, passed too
In particular stream segments are to be downloaded by the application,
decrypted (from passed keys if any), then re-assembled as appropriate:
we should expect TS segments, relying on ffmpeg is a solution to build a
proper MP4 file out of them.