Allow delegating HTTP fetch to native app
If we need to fetch some content, there are restrictions doing so from
the background script.
We can either try to find the tab and frame from which we can do what we
wish, or simply delegate the query to the native app, which allows us to
do the requests in similar conditions than the ones we intercepted.