ff: merge 'tabAdded' and 'tabCreated' events
Only retain 'tabCreated', since there is a true tab event for this, and
pass a 'created' field to indicate whether tab was truly just created,
or merely discovered on-the-fly.

Keep the 'tab' object in the handler, so that we can pass it to future
observers.
Not all fields are up-to-date, but this will do.